+++ core/integration/tests/server/purge_delete.rs
@@ -72,6 +72,12 @@ async fn should_delete_segments_with_consumer_group_barrier(harness: &TestHarnes
     partition.messages_required_to_save = "1",
     partition.enforce_fsync = "true",
 ))]
+// Consumer polls auto-commit offsets, so partition ops keep flowing while the
+// restarted node rejoins. Ops committed by the surviving quorum in that window
+// leave the pipeline before the rejoining replica can receive them; without
+// state transfer its journal gap is unfillable and the replica correctly
+// suicides ("replica is divergent") when the commit frontier crosses it.
+// Re-enable restart_on once state transfer lands.
 #[cfg_attr(not(feature = "vsr"), test_matrix([restart_off(), restart_on()]))]
 #[cfg_attr(feature = "vsr", test_matrix([restart_off()]))]
 async fn should_block_deletion_until_all_consumers_pass_segment(
@@ -89,12 +95,11 @@ async fn should_block_deletion_until_all_consumers_pass_segment(
 ))]
 // The scenario asserts the exact [0, 7, 14, 21] layout only on the legacy path;
 // under vsr it verifies the framing-agnostic purge outcome (offsets cleared,
-// files deleted, partition reset to a single segment at offset 0). The SDK
-// re-login panic is fixed, but restart_on stays vsr-gated on a separate
-// consumer-group blocker: after restart the heartbeat's group-assignment
-// refresh reconnect path re-binds an already-bound VSR session and fails
-// `AlreadyAuthenticated` (the consumerless `should_delete_segments_without_consumers`
-// restart_on passes, so this is CG-reconnect specific, not the login re-arm).
+// files deleted, partition reset to a single segment at offset 0). restart_on
+// is vsr-gated on the same rejoin-window state-transfer gap as
+// `should_block_deletion_until_all_consumers_pass_segment` above (the old
+// CG-reconnect `AlreadyAuthenticated` blocker is fixed by the SDK's
+// fresh-session reconnect).
 #[cfg_attr(not(feature = "vsr"), test_matrix([restart_off(), restart_on()]))]
 #[cfg_attr(feature = "vsr", test_matrix([restart_off()]))]
 async fn should_purge_topic_and_clear_consumer_offsets(
